Add files via upload
I tried to redo the files for the finding factors computation, we can work with this for now and fix it. Also I forgot to incorporate the @Test but I don't think I explicitly made any methods, at least not in the prototype class. So we still do need to check that since everything is kind of blank.