fixing the use of DataStores gRpc
ok i think I fixed the DataStore gRpc. Before the output was getting written in two places, the FactorComputationCoordinatorImpl  and the FactorServiceImpl, so i took what i had in the Service Impl and put it in the CoordinatorImpl. I still kept the original declaration of our ACTUAL FactorDataStoreImpl, just now we have an instance of the gRpc DataStore called dataStoreStub. I tried to keep a lot of what we had in the original code while also trying to implement the stub stuff for the server and client stuff. I also updated the FactorComputeEngineImpl since we forgot to add the delimiter to the formattedFactors output.